feat(packaging): prevent user config changes from being overwritten
Adds support for Debian "conffiles" to the packaging script.

When a script is packaged with an associated configuration file, the script now creates a `DEBIAN/conffiles` file within the package structure. This file lists the absolute path to the installed configuration file.

By marking the file as a conffile, `dpkg` will intelligently handle upgrades. If a user has modified their local version of the config, `dpkg` will prompt them to either keep their changes or accept the new version from the package, preventing accidental data loss.